Add files via upload
I changed the way the sky is partitioned. If the partition does not result in an equal integer number of pixels per part, the excess pixels will be distributed one for each of the pixels. For example, if we choose Nside = 64, there will be 49152 pixels. And if we choose Nparts = 1000, there were Npix / Nparts = 64 + rest, with rest = 152 pixels. These will be distributed in the first 152 parts. That is, the first part will not have 64, but 65 parts. Like the second part, the third, ...